making the chatbot better
